Use EXPECT_FALSE instead of EXPECT_EQ(false, X) (#752)
This fixes the failures we're finding from the CI builds. We also take the opportunity to improve the build system, to use ccache if it's available, and also to simplify the travis config. Now that we can just test with Boost 1.55 in the meantime, let's set the appropriate flags that will make the clang build/support work.

+++ boost/network/protocol/http/client/connection/async_normal.hpp
@@ -147,7 +147,7 @@ struct http_async_connection
       connection_delegate_ptr;
 
   http_async_connection(resolver_type& resolver, resolve_function resolve,
-                        bool follow_redirect, int timeout,
+                        bool follow_redirect, int64_t timeout,
                         bool remove_chunk_markers,
                         connection_delegate_ptr delegate)
       : timeout_(timeout),
@@ -578,7 +578,7 @@ struct http_async_connection
     }
   }
 
-  int timeout_;
+  int64_t timeout_;
   bool remove_chunk_markers_;
   boost::asio::steady_timer timer_;
   bool is_timedout_;
